Installers remove phpMyAdmin setup folder
The last version of ZPanel was retagged to include the phpMyAdmin setup
folder (which can be used by anyone to muck up your install) so the
upgrade scripts now delete it. Added it to the install scripts as well
in case someone forgets to delete it when updating the repo.